+++ src/sat/cryptominisat.h
@@ -15,7 +15,7 @@
 #ifdef BZLA_USE_CMS
 /*------------------------------------------------------------------------*/
 
-#include <cryptominisat5/cryptominisat.h>
+#include <cryptominisat.h>
 
 #include <memory>
 
@@ -39,11 +39,6 @@ class CryptoMiniSat : public SatSolver
   const char *get_version() const override;
 
   void set_num_threads(uint32_t n_threads) const;
-  /**
-   * Add new variable.
-   * @return The number of variables in the solver.
-   */
-  int32_t new_var() const;
 
  private:
   /**