+++ test/unit/api/test_capi.cpp
@@ -568,6 +568,14 @@ TEST_F(TestCApi, set_option)
     ASSERT_EQ(
         std::string(bitwuzla_get_option_mode(options, BITWUZLA_OPT_SAT_SOLVER)),
         "cadical");
+    bitwuzla_set_option_mode(options, BITWUZLA_OPT_SAT_SOLVER, "kissat");
+    ASSERT_EQ(
+        std::string(bitwuzla_get_option_mode(options, BITWUZLA_OPT_SAT_SOLVER)),
+        "kissat");
+    bitwuzla_set_option_mode(options, BITWUZLA_OPT_SAT_SOLVER, "cms");
+    ASSERT_EQ(
+        std::string(bitwuzla_get_option_mode(options, BITWUZLA_OPT_SAT_SOLVER)),
+        "cms");
     ASSERT_DEATH(
         bitwuzla_set_option_mode(options, BITWUZLA_OPT_BV_SOLVER, "asdf"),
         "invalid mode for option");